Move commands into internal directory
